Move EndianAware to own module
