+++ src/game_interface/dolphin/data_member.rs
@@ -3,7 +3,7 @@ use std::borrow::Cow;
 use bytemuck::CheckedBitPattern;
 use process_memory::{Architecture, CopyAddress, Memory, ProcessHandle, PutAddress};
 
-use crate::game_state::{GameMode, GameOstrich, GameState};
+use crate::endian::EndianAware;
 
 const GCN_BASE_ADDRESS: usize = 0x80000000;
 
@@ -114,25 +114,4 @@ impl<T: CheckedBitPattern + EndianAware> CheckedMemory<T> for DataMember<T> {
             .map_err(|_| std::io::ErrorKind::InvalidData)?;
         Ok(*val)
     }
-}
-
-pub trait EndianAware {
-    const NEEDS_SWAP: bool;
-}
-
-macro_rules! impl_aware {
-    ( false: $( $f:ty ),*; true: $( $t:ty ),* )=> {
-        $( impl_aware!(false, $f); )*
-        $( impl_aware!(true,  $t); )*
-    };
-    ( $needs_swap:expr, $typ:ty) => {
-        impl EndianAware for $typ {
-            const NEEDS_SWAP: bool = $needs_swap;
-        }
-    };
-}
-
-impl_aware!(false: bool, u8, GameMode, GameState, GameOstrich; true: i16, u32);
-impl<const N: usize> EndianAware for [u8; N] {
-    const NEEDS_SWAP: bool = false;
 }
